Refactor: removed src
